+++ src/ocrd/lib.bash
@@ -1,4 +1,0 @@
-echo "ERROR"
-echo "ERROR bashlib processors have been removed from OCR-D/core, use the Python API instead" >&2
-echo "ERROR"
-return 1