+++ src/ocrd_network/__init__.py
@@ -1,7 +1,6 @@
 from .client import Client
-from .constants import AgentType, JobState
+from .constants import JobState
 from .processing_server import ProcessingServer
 from .processing_worker import ProcessingWorker
-from .processor_server import ProcessorServer
 from .param_validators import DatabaseParamType, ServerAddressParamType, QueueServerParamType
 from .server_cache import CacheLockedPages, CacheProcessingRequests