+++ src/ocrd_network/models/__init__.py
@@ -10,13 +10,11 @@ __all__ = [
     'DBWorkflowScript',
     'PYJobInput',
     'PYJobOutput',
-    'PYOcrdTool',
     'PYResultMessage',
     'PYWorkflowJobOutput'
 ]
 
 from .job import DBProcessorJob, DBWorkflowJob, PYJobInput, PYJobOutput, PYWorkflowJobOutput
 from .messages import PYResultMessage
-from .ocrd_tool import PYOcrdTool
 from .workspace import DBWorkspace
 from .workflow import DBWorkflowScript