+++ src/ocrd_network/runtime_data/__init__.py
@@ -5,10 +5,9 @@ __all__ = [
     "DataNetworkAgent",
     "DataRabbitMQ",
     "DataProcessingWorker",
-    "DataProcessorServer"
 ]
 
 from .deployer import Deployer
 from .hosts import DataHost
-from .network_agents import DataNetworkAgent, DataProcessingWorker, DataProcessorServer
+from .network_agents import DataNetworkAgent, DataProcessingWorker
 from .network_services import DataMongoDB, DataRabbitMQ