+++ CHANGELOG.md
@@ -9,6 +9,11 @@ Changed:
 
   * `ocrd network client workflow run`: Allow passing workflow as `ocrd process` tasks as alternative to a workflow file, #1264, #1335
   * Processing Server: Store page ranges as comma-separated page IDs instead of relying on `..` numeric range operator, #1333, #1334
+  * `ocrd network`: More verbose error messages instead of plain assertions, #1337
+
+Removed:
+
+  * :fire: Remove Processor Server in favor of Processing Server + Processing Workers only, #1338, OCR-D/spec#265
 
 ## [3.5.1] - 2025-07-04